v4.2 AI changes
Specific sections of the status effect healing AI have been reverted to their original instance. Status effect healing, while not nonfunctional by any means, is having some unintended behaviors with certain skills (namely Charmdi) and item priority. This is an initial change that may be improved in the future.

Twilight’s conditional no longer checks for Evasion down, in line with its new effect.

Acquiesce’s conditionals have been modified to increase the likelihood of the move being used in certain situations.
Act Freely and Healing and Support:
If ( under 10% SP ) to ( under 15% SP )
If ( under 15% SP AND 80% HP ) to ( 30% SP and 80% HP )
Removed: If ( under 20% SP AND 70% HP )
Conserve SP:
If ( under 35% SP AND 80% HP ) to ( 50% SP and 80% HP )
Removed: If ( under 40% SP AND 70% HP )